Random Question Generator and Answer Bot
I've been studying Javascript for a week or two. Told myself only yesterday that I was gonna keep things simple and not get ahead of myself. Disregarded that completely today and made this. At the moment, there seems to be a bug somewhere in the code. For some reason, when the question "Am I good looking?" is displayed, the answerBot returns nothing. I've no idea where to start looking and don't even know if the different programs in my code are written in the correct order. Quite chuffed nonetheless. :-)